I was not happy with the generated assembly for these functions.
They were needless spilling registers and ensuring stack aligment. So I wrote it in assembly instead. Not like I would expect the compiler to be able to know that  bunki_ctx_stack_start only clobbers the RAX register and does not care about stack aligment. So it was doing the correct thing in the general case and following standard calling convention.