Was still getting random crashes. Just realized I was trying to write an illegal MXCSR value.
When testing calls to a coroutine that has returned permently I had random crashes. I realized I was trying to write garbage to the MXCSR register(since save and load modify it). Some values are illegal such as reserved bit ect...  So make sure to save it so I don't try and write garbage.